Refactoring input handling
Given that only one sample is supported and the the channel
unpacking and repacking code is become increasingly cumbersome
and verbose, I've moved all parameters into params and removed
the samplesheet inputs, streamlining and simplifying the down-
stream code. It will now be much less of a pain to add new
parameters.